requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
streamlit
python-pptx
reportlab
Flask
pycaret
gunicorn
